# compute: better-fueled linear join