--- a/src/compute-client/src/protocol/command.proto
+++ b/src/compute-client/src/protocol/command.proto
@@ -65,4 +65,5 @@ message ProtoComputeParameters {
     optional uint32 max_result_size = 1;
     mz_persist_client.cfg.ProtoPersistParameters persist = 2;
     optional uint64 dataflow_max_inflight_bytes = 3;
+    optional bool enable_mz_join_core = 4;
 }