--- a/src/compute/src/render/join/mod.rs
+++ b/src/compute/src/render/join/mod.rs
@@ -13,5 +13,6 @@
 
 mod delta_join;
 mod linear_join;
+mod mz_join_core;
 
 pub use linear_join::LinearJoinImpl;