… src/compute/src/render/join/mz_join_core.rs
… src/sql/src/session/vars.rs
… test/sqllogictest/joins.slt
… test/testdrive/linear-join-fuel.td