This PR changes the way linear joins are fueled. Rather than checking the remaining fuel only between keys, the new approach checks every time all matches for a batch value have been produced.

Doing so enables the join operator to regularly yield even when computing joins that have only a single or few keys. This should lead to much better responsiveness of replicas that have dataflows with cross-joins or skewed joins installed.

As this fueling change requires more control over the internals of the join operator than the DD API affords us, part (really: the bulk) of this PR is concerned with copying the join operator code from DD into the MZ codebase. In doing so, it also significantly simplifies the original join operator code, by removing logic and genericity not required by MZ. In general, no attempt is made at keeping the compute join implementation close to the one in DD. This will make it harder to upstream changes, should we ever want to do so, but at the same time gives us the opportunity to make the code more readable. I anticipate that we will want to perform more tuning on the join code in the future (e.g. by adding a shutdown token), so making it easy to modify seems more important than attempting compatibility with DD.

See performance test results here: https://github.com/MaterializeInc/materialize/pull/19186#issuecomment-1540078479

### Proof of Efficacy

To test that this actually has the desired effect of making the join operator more responsive, we can run a test with a cross join:
```sql
CREATE TABLE t (a int);
INSERT INTO t SELECT generate_series(1, 10000);
SELECT sum(t1.a + t2.a * 10000) FROM t t1, t t2;

-- in a second session:
\timing
\o /dev/null
SELECT * FROM mz_internal.mz_dataflow_operators;
SELECT * FROM mz_internal.mz_dataflow_operators;
SELECT * FROM mz_internal.mz_dataflow_operators;
```

On main, this yields the timings on my laptop:
```
Time: 39232.076 ms (00:39.232)
Time: 239.962 ms
Time: 6.080 ms
```

The first query only returns after the cross join has returned (or roughly at the same time). That's because the join operator never yields, so the replica can't process commands while it is running.

On this branch, these timings are printed:
```
Time: 538.751 ms
Time: 872.802 ms
Time: 865.107 ms
```

Queries return promptly even as the cross join is running. The experience is a bit sluggish, but I think this can be solved with a different (e.g. time-based) fueling strategy. I consider that out-of-scope for this PR though.

### Motivation

  * This PR fixes a recognized bug.

Fixes #19093.

### Tips for reviewer

There are two ways to read this PR:

1. Read the full diff. It's not huge, but perhaps overwhelming if you are not familiar with the DD join operator. Most of the positive changes are taken verbatim from the DD implementation and it's not easy to make out what the changes are. So I would generally recommend you go with option (2).
2. Look at the commits separately. I've tried to structure them in a way that makes it obvious what changes are made from the DD join implementation. The first commit simply copies over the relevant code from DD without changes, the subsequent ones modify it in various ways (applying rustfmt, fixing clippy warnings, simplifying). The second-to-last commit introduces the change in fueling. Look at the commit messages for more details.

~~Do you think this change warrants being rolled out with a feature flag? Personally I think that's not necessary, as we don't actually introduce new logic, just modify what's already existing. But I'd be fine with adding a feature flag if folk would like to move slower with this.~~ We agreed on adding a feature flag because there are (unlikely) workloads for this new join implementation is considerably slower than the old one and we want to option of switching back quickly. 

Also let me know if you think I should run other performance tests to ensure there are no regressions!

### Checklist

- [x] This PR has adequate test coverage / QA involvement has been duly considered.
- [x] This PR has an associated up-to-date [design doc](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/README.md), is a design doc ([template](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/design/00000000_template.md)), or is sufficiently small to not require a design.
  <!-- Reference the design in the description. -->
- [x] This PR evolves [an existing `$T ⇔ Proto$T` mapping](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/command-and-response-binary-encoding.md) (possibly in a backwards-incompatible way) and therefore is tagged with a `T-proto` label.
- [x] If this PR will require changes to cloud orchestration, there is a companion cloud PR to account for those changes that is tagged with the release-blocker label ([example](https://github.com/MaterializeInc/cloud/pull/5021)).
  <!-- Ask in #team-cloud on Slack if you need help preparing the cloud PR. -->
- [x] This PR includes the following [user-facing behavior changes](https://github.com/MaterializeInc/materialize/blob/main/doc/developer/guide-changes.md#what-changes-require-a-release-note):
  - Make replicas that serve dataflows with cross joins or joins with only few keys stay responsive to user queries.