@teskje 2023-05-10
@philip-stoev This touches a fundamental part of compute, so I think having more eyes on it is warranted. I'm starting a nightly test run now, but lmk if there is anything else we can/should test.