@teskje 2023-05-10
Nightlies are green.

The coverage run revealed a couple places that are not covered in the new code, so I added another commit adding tests to fix this. New coverage run: https://buildkite.com/materialize/coverage/builds/96