@teskje 2023-05-10
Thanks for the review @vmarcos!

1. My assumption is indeed that the DD join operator will not change much in the future. But also a) getting changes upstreamed will be difficult due to Frank's limited bandwidth, b) scoping the implementation to MZ means a significant reduction of complexity, as we can make a lot of assumptions that DD code can't, and c) the extent of the operator is small enough that it seems feasible to keep up with changes in DD even if that involves translating them to the diverged MZ implementation. That said, I appreciate that this might be a controversial stance, so happy to hear other opinions!
2. Sure, I will run the benchmarks with multiple workers/processes too!
3. Would the cross-join suffice as a skewed workload? Or were you thinking of something not _that_ skewed?