@teskje 2023-05-11
> I pushed a .td test based on your original manual test that confirms that yielding indeed has the desired end-to-end result. 

This is awesome, thank you!