@teskje 2023-05-11
## Performance Tests -- Multi-worker

I ran the same performance tests from above again in a multi-worker setup. Specifically, I made the following changes:

* Use a `medium` replica, rather than an `xsmall`.
* Run TPC-H with scale-factor 10 rather than 1.

These are the results:

### Cross-join

Results [ms]:

main | this branch | ratio
---: | ---: | ---:
77585 | 69070 | 0.89

This is very similar to the single-worker result, which makes sense since a cross-join only has a single key, so pretty much all the processing only happens on a single worker anyway.

### TPC-H

Results [ms]:

query | main | this branch | ratio
:--- | ---: | ---: | ---:
Q01 | 57304 | 53363 | 0.93
Q02 | 7878 | 7997 | 1.02
Q03 | 48020 | 39439 | 0.82
Q04 | 45411 | 41544 | 0.91
Q05 | 47072 | 47225 | 1.0
Q06 | 47719 | 35978 | 0.75
Q07 | 46390 | 48946 | 1.06
Q08 | 52835 | 51301 | 0.97
Q09 | 67074 | 69738 | 1.04
Q10 | 48232 | 50534 | 1.05
Q11 | 6261 | 5927 | 0.95
Q12 | 36923 | 37267 | 1.01
Q13 | 16386 | 15775 | 0.96
Q14 | 39009 | 33778 | 0.87
Q15 | 35124 | 34621 | 0.99
Q16 | 6678 | 7319 | 1.1
Q17 | 52040 | 45285 | 0.87
Q18 | 73589 | 67643 | 0.92
Q19 | 40166 | 38636 | 0.96
Q20 | 375575 | 339703 | 0.9
Q21 | 59490 | 55717 | 0.94
Q22 | 14235 | 12261 | 0.86

This shows a lot more variability than the single-worker case. Measurements generally look to be in favor of the new join implementation, but I wouldn't trust them too much. For example, Q01 doesn't even contain a join and shows 7% difference. I also noticed quite a large variability between benchmark runs on the same MZ build, so I assume the numbers would even out over more runs. Still the measurements seem generally in favor of the new implementation too.