@teskje 2023-05-12
## Performance Tests -- Multi-worker -- Skewed Join

And here is the promised test using a skewed join workload.

I created a skewed data distribution using this simple Python code (credits go to @tokenrove):
```python
import random
from collections import Counter

counts = Counter(random.randint(0, 2**64-1).bit_length() for _ in range(20000))
for k, n in enumerate(counts.values()):
    print(f"INSERT INTO t SELECT {k}, generate_series(1, {n});")
```

... then used the output to run this experiment:
```sql
CREATE TABLE t (a int, b int);
INSERT INTO t SELECT 0, generate_series(1, 9924);
INSERT INTO t SELECT 1, generate_series(1, 5064);
INSERT INTO t SELECT 2, generate_series(1, 2557);
INSERT INTO t SELECT 3, generate_series(1, 157);
INSERT INTO t SELECT 4, generate_series(1, 1271);
INSERT INTO t SELECT 5, generate_series(1, 584);
INSERT INTO t SELECT 6, generate_series(1, 73);
INSERT INTO t SELECT 7, generate_series(1, 291);
INSERT INTO t SELECT 8, generate_series(1, 1);
INSERT INTO t SELECT 9, generate_series(1, 43);
INSERT INTO t SELECT 10, generate_series(1, 10);
INSERT INTO t SELECT 11, generate_series(1, 14);
INSERT INTO t SELECT 12, generate_series(1, 8);
INSERT INTO t SELECT 13, generate_series(1, 3);

\timing
SELECT sum(t1.b + t2.b * 10000) FROM t t1 JOIN t t2 USING (a); 
```

The experiment was again run on a `medium`-size replica. This size uses 12 workers, so each worker gets _some_ processing to do.

Results (best of three) [ms]:

<!DOCTYPE html>

| main | this branch | ratio |
| ---: | ---: | ---: |
| 100343 | 78881 | 0.79 |

The performance wins here seem to be more significant than for the cross-join. We should probably again be wary of reading too much into the results considering the variability seen with TPC-H, but I think it's save to say that we don't need to expect regressions.