@teskje 2023-05-12
> I am assuming that even if you are running a multi-worker skewed join, you can of course run other queries without losing interactivity.

Yes, these are the query times from the above skewed-join workload:

```sql
materialize=> \timing
Timing is on.
materialize=> \o /dev/null
materialize=> select * from mz_internal.mz_dataflow_operators;
select * from mz_internal.mz_dataflow_operators;
select * from mz_internal.mz_dataflow_operators;
Time: 1225.766 ms (00:01.226)
Time: 1443.466 ms (00:01.443)
Time: 1446.125 ms (00:01.446)
```

It's not snappy but still a huge improvement over potentially infinite query times :)