@teskje 2023-05-16
@antiguru pointed out a stray `#[inline(never)]` in the join code that probably doesn't make sense to have. I removed it and it turns out to be great for performance!

These are comparisons between the second-to-last commit (with disallowed inlining) and the last commit, using the performance tests from above:

### Cross-join

before | now | ratio
---: | ---: | ---:
69070 | 66629 | 0.96 

### Skewed join

before | now | ratio
---: | ---: | ---:
78881 | 68986 | 0.87 

### TPC-H

query | before | now | ration
:--- | ---: | ---: | ---:
Q01 | 66974 | 53363 | 0.80
Q02 | 13377 | 7997 | 0.60
Q03 | 46605 | 39439 | 0.85
Q04 | 49483 | 41544 | 0.84
Q05 | 57777 | 45670 | 0.79
Q06 | 39902 | 35978 | 0.90
Q07 | 61234 | 48946 | 0.80
Q08 | 58986 | 51301 | 0.87
Q09 | 68018 | 69738 | 1.03
Q10 | 69203 | 50534 | 0.73
Q11 | 11340 | 5927 | 0.52
Q12 | 52928 | 37267 | 0.70
Q13 | 17936 | 15775 | 0.88
Q14 | 37374 | 33485 | 0.90
Q15 | 35227 | 34621 | 0.98
Q16 | 7073 | 7319 | 1.03
Q17 | 58866 | 45285 | 0.77
Q18 | 75762 | 66622 | 0.88
Q19 | 47748 | 37520 | 0.79
Q20 | 339554 | 326566 | 0.96
Q21 | 107615 | 55717 | 0.52
Q22 | 14880 | 12251 | 0.82

As always, caveats apply due to the high variance of especially the TPC-H results.